Fri, Dec 30, 2022, 8:43 AM +00:00

+++ sentimental-credit/credit.py
@@ -11,7 +11,7 @@ def main():
     while(number > 0):
         digit = number % 10
         sum = sum + digit
-        number = number / 100
+        number = int(number / 100)
     print(sum)
